compiler warnings with Qt5 fixed
git-svn-id: svn://svn.code.sf.net/p/qwt/code/trunk/qwt@1582 29e17f8d-cf37-471e-a288-72cc03a76eda

+++ src/qwt_event_pattern.cpp
@@ -217,7 +217,7 @@ bool QwtEventPattern::mouseMatch( const MousePattern &pattern,
 
     const bool matched =
         ( event->modifiers() & Qt::KeyboardModifierMask ) ==
-            ( int )( pattern.state & Qt::KeyboardModifierMask );
+            ( pattern.state & Qt::KeyboardModifierMask );
 
     return matched;
 }
@@ -268,7 +268,7 @@ bool QwtEventPattern::keyMatch(
 
     const bool matched =
         ( event->modifiers() & Qt::KeyboardModifierMask ) ==
-            ( int )( pattern.state & Qt::KeyboardModifierMask );
+            ( pattern.state & Qt::KeyboardModifierMask );
 
     return matched;
 }
